refact: AWS SDKを使用箇所をRedisMessageProcessorに移動し、リファクタ
